% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liver_plot.R
\name{liver_plot}
\alias{liver_plot}
\title{liver_plot}
\usage{
liver_plot(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-08-20"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
liver_plot
}
